modules/create-location-grpc/requirements.txt
grpcio==1.41.0 
grpcio-tools==1.41.0
DateTime==4.3
urllib3==1.26.7
kafka-python==2.0.2